rename hubble-proxy to hubble-relay
The scope of this component goes beyond simply proxying requests.
As such, it was decided to rename it to "hubble-relay". This is what
this commit does, without any functional changes.
