--- hubble-proxy/.gitignore
@@ -1,1 +1,0 @@
-hubble-proxy
